feat: 주류 상세정보 페이지 디자인 수정 및 컴포넌트 경로 리팩토링
